--- coordinator/tests/integration/tester-grpc/charmcraft.yaml
@@ -1,17 +1,0 @@
-# Learn more about charmcraft.yaml configuration at:
-# https://juju.is/docs/sdk/charmcraft-config
-type: "charm"
-bases:
-  - build-on:
-    - name: "ubuntu"
-      channel: "22.04"
-    run-on:
-    - name: "ubuntu"
-      channel: "22.04"
-parts:
-  charm:
-    charm-binary-python-packages:
-      - "pydantic>=2"
-      - "cryptography"
-      - "jsonschema"
-      - "opentelemetry-exporter-otlp-proto-http==1.21.0"
