--- coordinator/tests/integration/tester-grpc/requirements.txt
@@ -1,2 +1,0 @@
-ops >= 1.4.0
-opentelemetry-exporter-otlp-proto-grpc>=1.21.0
