--- coordinator/tests/integration/tester/README.md
@@ -1,2 +1,0 @@
-to deploy:
-`juju deploy /path/to/tester.charm --resource workload=python:slim-buster --config tempo_endpoint:foo.bar:4318`
