--- coordinator/tests/integration/tester/requirements.txt
@@ -1,1 +1,0 @@
-ops >= 1.4.0
