--- coordinator/tests/integration/tester/src/resources/webserver-dependencies.txt
@@ -1,7 +1,0 @@
-requests
-opentelemetry-exporter-otlp
-opentelemetry-instrumentation-fastapi
-opentelemetry-instrumentation-logging
-fastapi
-uvicorn
-prometheus_client
